+++ src/Concerto/PanelBundle/Resources/starter_content/src/template/inputSlider/inputSlider.js
@@ -1,0 +1,26 @@
+testRunner.compileProvider.component('inputSlider', {
+  templateUrl: testRunner.settings.platformUrl + "ViewTemplate/inputSlider/content?css=1,html=1",
+  bindings: {
+    field: '=',
+    values: '='
+  },
+  controller: function controller($scope) {
+    this.$onInit = function() {
+      $scope.field = this.field;
+      $scope.values = this.values;
+      
+      console.log($scope.getValidator('max').value);
+    };
+
+    $scope.getValidator = function (type) {
+      if (typeof this.field.validation === 'undefined') return null;
+      for (var i = 0; i < this.field.validation.length; i++) {
+        var validation = this.field.validation[i];
+        if (validation.type === type) {
+          return validation;
+        }
+      }
+      return null;
+    };
+  }
+});